requirements.txt
Jinja2-2.8 
MarkupSafe-0.23 
Werkzeug-0.10.4 
flask-0.10.1 
itsdangerous-0.24
pymysql-0.6.6
